Adds build specific logic
If the modules are required at build time a seperate layer will be
created that will install node_modules with the same logic as setting
NODE_ENV=development meaning all devDependencies will be available
for subsequent buildpacks to access. However only dependencies that
conform to the defaut or user set NODE_ENV will make it onto the final
running image.